refactor: 네트워크 모듈 구조 정리 (#37)
